+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ## mini.starter
 
+- Allow `config.header` and `config.footer` be any value, which will be converted to string via `tostring()`.
 - Update query logic to not allow queries which result into no items.